Add action hooks to provide better compatibility with other plugins.
The do_action() hooks allow site owners to write custom code that
triggers upon various WP-Polls events. This is useful when, for example,
a site owner is running a caching plugin that needs to be told to clear
the cache for a given page when a poll on that page is updated.

See also: http://codex.wordpress.org/Function_Reference/do_action