preventing cross site voting & validating answers
There is a current flaw in wp-polls that allows people to embed the following into a page:
http://example.com/wp-admin/admin-ajax.php?action=polls&view=process&poll_id=POLLID&poll_POLLID=ANSWER&poll_POLLID_nonce=POLLNONCE

This allows exploiters to get votes from people who have not even visited the WordPress site. I exploited this on a WordPress site, and got 318k votes by embedding the link into various forum signatures. However, since the poll answer is a GET variable, the actual answers went to 0%. It is because the votes were not counted as answers, but still added to the total votes. I added an if statement that checks if the poll answer is legitimate. This prevents fraudulent answers, and cross site voting.